function that prints the binary representation of a number.

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -7,7 +7,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned long int mask = (unsigned long int)1 << (sizeof(unsigned long int) * 8 - 1);
+	unsigned long int mask = 1UL << (sizeof(unsigned long int) * 8 - 1);
 	int leading_zeroes = 1;
 
 	while (mask != 0)
